ci(psalm): fix wrong PHP version

+++ app/Http/Controllers/MovieCollectionController.php
@@ -5,16 +5,18 @@ declare(strict_types=1);
 namespace App\Http\Controllers;
 
 use App\Models\Movie;
+use App\Models\User;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Http\RedirectResponse;
 
 class MovieCollectionController extends Controller
 {
-    public function store(Movie $movie)
+    public function store(Movie $movie): RedirectResponse
     {
-        // If logged in, add to favorites
-        if (Auth::check()) {
-            $user = Auth::user();
+        /** @var User|null */
+        $user = Auth::user();
 
+        if (Auth::check() && $user !== null) {
             $user->collection()->attach($movie);
 
             return back();
@@ -23,12 +25,12 @@ class MovieCollectionController extends Controller
         return back();
     }
 
-    public function destroy(Movie $movie)
+    public function destroy(Movie $movie): RedirectResponse
     {
-        // If logged in, remove from favorites
-        if (Auth::check()) {
-            $user = Auth::user();
+        /** @var User|null */
+        $user = Auth::user();
 
+        if (Auth::check() && $user !== null) {
             $user->collection()->detach($movie);
 
             return back();
